Move all the type check warning to a transform
