+++ lib/assets/Asset.js
@@ -183,19 +183,6 @@ class Asset extends EventEmitter {
     }
   }
 
-  _warnIncompatibleTypes(incompatibleTypes) {
-    this._issuedWarnings = this._issuedWarnings || {};
-    const errorString = `Asset is used as both ${[...incompatibleTypes]
-      .sort()
-      .join(' and ')}`;
-    if (!this._issuedWarnings[errorString]) {
-      this._issuedWarnings[errorString] = true;
-      const err = new Error(errorString);
-      err.asset = this;
-      this.assetGraph.warn(err);
-    }
-  }
-
   _inferType(incomingRelation) {
     if (!this._inferredType) {
       let typeFromContentType;
@@ -210,41 +197,25 @@ class Asset extends EventEmitter {
         this._inferredType === 'Image' ||
         this._inferredType === 'Font'
       ) {
-        let commonTargetType;
+        let firstFoundTargetType;
         const incomingRelations = this.incomingRelations;
         if (incomingRelation) {
           incomingRelations.push(incomingRelation);
         }
         for (const incomingRelation of incomingRelations) {
           if (incomingRelation.targetType) {
-            if (!commonTargetType) {
-              commonTargetType = incomingRelation.targetType;
-            } else if (
-              commonTargetType !== incomingRelation.targetType &&
-              !AssetGraph[commonTargetType].prototype[
-                `is${incomingRelation.targetType}`
-              ] &&
-              !AssetGraph[incomingRelation.targetType].prototype[
-                `is${commonTargetType}`
-              ]
-            ) {
-              this._warnIncompatibleTypes([
-                commonTargetType,
-                incomingRelation.targetType
-              ]);
-              commonTargetType = undefined;
-              break;
-            }
+            firstFoundTargetType = incomingRelation.targetType;
+            break;
           }
         }
-        if (commonTargetType) {
+        if (firstFoundTargetType) {
           if (
-            this._inferredType &&
-            !AssetGraph[commonTargetType].prototype[`is${this._inferredType}`]
+            !this._inferredType ||
+            AssetGraph[firstFoundTargetType].prototype[
+              `is${this._inferredType}`
+            ]
           ) {
-            this._warnIncompatibleTypes([this._inferredType, commonTargetType]);
-          } else {
-            this._inferredType = commonTargetType;
+            this._inferredType = firstFoundTargetType;
           }
         }
       }
@@ -304,23 +275,6 @@ class Asset extends EventEmitter {
             break;
           }
           superclass = Object.getPrototypeOf(superclass);
-        }
-        if (this.location === undefined) {
-          if (!this._isCompatibleWith(type)) {
-            this._warnIncompatibleTypes([this.type, type]);
-          } else {
-            const incompatibleIncomingTargetTypes = this.incomingRelations.filter(
-              incomingRelation =>
-                incomingRelation.targetType &&
-                !this._isCompatibleWith(incomingRelation.targetType)
-            );
-            if (incompatibleIncomingTargetTypes.length) {
-              this._warnIncompatibleTypes([
-                this.type,
-                ...incompatibleIncomingTargetTypes.map(i => i.targetType)
-              ]);
-            }
-          }
         }
       }
     }
@@ -452,17 +406,7 @@ class Asset extends EventEmitter {
               /^\s*([\w\-+.]+\/[\w-+.]+)(?:\s|;|$)/i
             );
             if (matchContentType) {
-              const contentType = matchContentType[1].toLowerCase();
-              const typeFromContentType =
-                AssetGraph.typeByContentType[contentType];
-              if (
-                typeFromContentType &&
-                this.location === undefined &&
-                !this._isCompatibleWith(typeFromContentType, contentType)
-              ) {
-                this._warnIncompatibleTypes([this.type, typeFromContentType]);
-              }
-              this.contentType = contentType;
+              this.contentType = matchContentType[1].toLowerCase();
 
               const matchCharset = contentTypeHeaderValue.match(
                 /;\s*charset\s*=\s*(['"]|)\s*([\w-]+)\s*\1(?:\s|;|$)/i
@@ -1098,17 +1042,6 @@ class Asset extends EventEmitter {
             outgoingRelation.to,
             outgoingRelation
           );
-          const targetType = outgoingRelation.targetType;
-          if (
-            targetType &&
-            outgoingRelation.to.location === undefined &&
-            !outgoingRelation.to._isCompatibleWith(targetType)
-          ) {
-            outgoingRelation.to._warnIncompatibleTypes([
-              targetType,
-              outgoingRelation.to.type
-            ]);
-          }
         }
       }
     }
@@ -1501,26 +1434,6 @@ class Asset extends EventEmitter {
       }
     }
     return clone;
-  }
-
-  // contentType is optional, but can be passed when the determination of Class was based on a definitely known Content-Type
-  _isCompatibleWith(Class, contentType) {
-    if (typeof Class === 'undefined') {
-      Class = Asset;
-    } else if (typeof Class === 'string') {
-      Class = this.assetGraph[Class];
-    }
-    if (contentType === 'text/plain' && this.type && this.type !== 'Text') {
-      // Don't allow an isText asset to pass as compatible with explicit text/plain
-      return false;
-    }
-    return (
-      this instanceof Class ||
-      !this._type ||
-      Class.prototype instanceof AssetGraph[this._type] ||
-      !!(this.isImage && Class === AssetGraph.Image) || // Svg is modelled as a subclass of Xml, not Image
-      !!(this.isImage && Class === AssetGraph.Font) // Svg can be used as a font as well
-    );
   }
 
   /**
